requirements.txt
langgraph
httpx
ipython
graphviz
langchain
langchain-ollama
termcolor
rich
ollama
docker

torch
torchaudio
sounddevice
scipy
faster-whisper
lightning-whisper-mlx
kafka-python
pydantic
pydub
requests
PyYAML
confluent-kafka[avro]
